# Playwright Workspaces/apiDefinition.swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Playwright Workspaces",
        "description": "Full browser automation powered by Azure Playwright Workspaces remote browsers. Create sessions, navigate, click, type, scrape data, take screenshots, and automate multi-step web workflows — all running on managed cloud Chromium instances.",
        "version": "1.0.0",
        "contact": {
            "name": "Troy Taylor",
            "email": "troy@troystaylor.com"
        }
    },
    "host": "pw-workspaces-broker.azurewebsites.net",
    "basePath": "/api",
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/sessions/{sessionId}": {
            "get": {
                "operationId": "GetSession",
                "summary": "Get Session Status",
                "description": "Get the current status of a browser session including page URL, title, and time remaining.",
                "tags": [
                    "Sessions"
                ],
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "Session ID",
                        "x-ms-summary": "Session ID",
                        "x-ms-url-encoding": "single"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session details",
                        "schema": {
                            "$ref": "#/definitions/SessionStatusResponse"
                        }
                    },
                    "404": {
                        "description": "Session not found or expired"
                    }
                }
            },
            "delete": {
                "operationId": "CloseSession",
                "summary": "Close Session",
                "description": "Close a browser session and release the remote browser. Always close sessions when done to avoid unnecessary charges.",
                "tags": [
                    "Sessions"
                ],
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "Session ID",
                        "x-ms-summary": "Session ID",
                        "x-ms-url-encoding": "single"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Session closed"
                    },
                    "404": {
                        "description": "Session not found"
                    }
                }
            }
        },
        "/sessions/{sessionId}/navigate": {
            "post": {
                "operationId": "Navigate",
                "summary": "Navigate to URL",
                "description": "Navigate the remote browser to a new URL while preserving cookies and session state.",
                "tags": [
                    "Navigation"
                ],
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "Session ID",
                        "x-ms-summary": "Session ID",
                        "x-ms-url-encoding": "single"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": [
                                "url"
                            ],
                            "properties": {
                                "url": {
                                    "type": "string",
                                    "description": "URL to navigate to",
                                    "x-ms-summary": "URL"
                                },
                                "waitUntil": {
                                    "type": "string",
                                    "description": "When to consider navigation complete",
                                    "x-ms-summary": "Wait Until",
                                    "enum": [
                                        "load",
                                        "domcontentloaded",
                                        "networkidle"
                                    ],
                                    "default": "networkidle"
                                },
                                "waitForSelector": {
                                    "type": "string",
                                    "description": "CSS selector to wait for after navigation (optional)",
                                    "x-ms-summary": "Wait for Selector"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Navigation complete",
                        "schema": {
                            "$ref": "#/definitions/NavigationResponse"
                        }
                    },
                    "404": {
                        "description": "Session not found"
                    },
                    "408": {
                        "description": "Navigation timeout"
                    }
                }
            }
        },
        "/sessions/{sessionId}/click": {
            "post": {
                "operationId": "ClickElement",
                "summary": "Click Element",
                "description": "Click an element on the page by CSS selector, text content, or ARIA role.",
                "tags": [
                    "Interaction"
                ],
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "Session ID",
                        "x-ms-summary": "Session ID",
                        "x-ms-url-encoding": "single"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "selector": {
                                    "type": "string",
                                    "description": "CSS selector of element to click",
                                    "x-ms-summary": "CSS Selector"
                                },
                                "text": {
                                    "type": "string",
                                    "description": "Click element containing this text (alternative to selector)",
                                    "x-ms-summary": "Text Content"
                                },
                                "role": {
                                    "type": "string",
                                    "description": "ARIA role of element to click (e.g., button, link)",
                                    "x-ms-summary": "ARIA Role"
                                },
                                "roleName": {
                                    "type": "string",
                                    "description": "Accessible name when using role-based click",
                                    "x-ms-summary": "Role Name"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Element clicked",
                        "schema": {
                            "$ref": "#/definitions/ActionResponse"
                        }
                    },
                    "404": {
                        "description": "Session or element not found"
                    }
                }
            }
        },
        "/sessions/{sessionId}/type": {
            "post": {
                "operationId": "TypeText",
                "summary": "Type Text",
                "description": "Type text into an input field identified by CSS selector, label, or placeholder.",
                "tags": [
                    "Interaction"
                ],
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "Session ID",
                        "x-ms-summary": "Session ID",
                        "x-ms-url-encoding": "single"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": [
                                "text"
                            ],
                            "properties": {
                                "selector": {
                                    "type": "string",
                                    "description": "CSS selector of input element",
                                    "x-ms-summary": "CSS Selector"
                                },
                                "label": {
                                    "type": "string",
                                    "description": "Label text of input field (alternative to selector)",
                                    "x-ms-summary": "Field Label"
                                },
                                "placeholder": {
                                    "type": "string",
                                    "description": "Placeholder text of input field (alternative to selector)",
                                    "x-ms-summary": "Placeholder"
                                },
                                "text": {
                                    "type": "string",
                                    "description": "Text to type into the field",
                                    "x-ms-summary": "Text"
                                },
                                "submit": {
                                    "type": "boolean",
                                    "description": "Press Enter after typing (default: false)",
                                    "x-ms-summary": "Submit"
                                },
                                "clear": {
                                    "type": "boolean",
                                    "description": "Clear existing text before typing (default: true)",
                                    "x-ms-summary": "Clear First"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Text entered",
                        "schema": {
                            "$ref": "#/definitions/ActionResponse"
                        }
                    },
                    "404": {
                        "description": "Session or element not found"
                    }
                }
            }
        },
        "/sessions/{sessionId}/select": {
            "post": {
                "operationId": "SelectOption",
                "summary": "Select Option",
                "description": "Select an option from a dropdown or select element.",
                "tags": [
                    "Interaction"
                ],
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "Session ID",
                        "x-ms-summary": "Session ID",
                        "x-ms-url-encoding": "single"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": [
                                "selector"
                            ],
                            "properties": {
                                "selector": {
                                    "type": "string",
                                    "description": "CSS selector of the select element",
                                    "x-ms-summary": "Selector"
                                },
                                "value": {
                                    "type": "string",
                                    "description": "Option value to select",
                                    "x-ms-summary": "Value"
                                },
                                "label": {
                                    "type": "string",
                                    "description": "Option label/text to select (alternative to value)",
                                    "x-ms-summary": "Label"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Option selected",
                        "schema": {
                            "$ref": "#/definitions/ActionResponse"
                        }
                    },
                    "404": {
                        "description": "Session or element not found"
                    }
                }
            }
        },
        "/sessions/{sessionId}/fill-form": {
            "post": {
                "operationId": "FillForm",
                "summary": "Fill Form",
                "description": "Fill multiple form fields at once using label-to-value or selector-to-value mappings.",
                "tags": [
                    "Interaction"
                ],
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "Session ID",
                        "x-ms-summary": "Session ID",
                        "x-ms-url-encoding": "single"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": [
                                "fields"
                            ],
                            "properties": {
                                "fields": {
                                    "type": "object",
                                    "description": "Map of field selectors or labels to values (e.g., {\"#email\": \"user@example.com\", \"Password\": \"***\"})",
                                    "x-ms-summary": "Fields"
                                },
                                "submit": {
                                    "type": "boolean",
                                    "description": "Click submit button after filling (default: false)",
                                    "x-ms-summary": "Submit Form"
                                },
                                "submitSelector": {
                                    "type": "string",
                                    "description": "CSS selector of submit button (default: auto-detect)",
                                    "x-ms-summary": "Submit Button Selector"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Form filled",
                        "schema": {
                            "$ref": "#/definitions/FormFillResponse"
                        }
                    },
                    "404": {
                        "description": "Session not found"
                    }
                }
            }
        },
...
        "/sessions/{sessionId}/screenshot": {
            "get": {
                "operationId": "TakeScreenshot",
                "summary": "Take Screenshot",
                "description": "Capture a screenshot of the current page in the remote browser.",
                "tags": [
                    "Utility"
                ],
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "Session ID",
                        "x-ms-summary": "Session ID",
                        "x-ms-url-encoding": "single"
                    },
                    {
                        "name": "fullPage",
                        "in": "query",
                        "type": "boolean",
                        "description": "Capture full scrollable page (default: false)",
                        "x-ms-summary": "Full Page"
                    },
                    {
                        "name": "format",
                        "in": "query",
                        "type": "string",
                        "enum": [
                            "png",
                            "jpeg"
                        ],
                        "description": "Image format (default: png)",
                        "x-ms-summary": "Format"
                    },
                    {
                        "name": "selector",
                        "in": "query",
                        "type": "string",
                        "description": "CSS selector to screenshot a specific element",
                        "x-ms-summary": "Element Selector"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Screenshot captured",
                        "schema": {
                            "$ref": "#/definitions/ScreenshotResponse"
                        }
                    },
                    "404": {
                        "description": "Session not found"
                    }
                }
            }
        },
        "/sessions/{sessionId}/evaluate": {
            "post": {
                "operationId": "EvaluateJavaScript",
                "summary": "Evaluate JavaScript",
                "description": "Execute JavaScript code on the page and return the result.",
                "tags": [
                    "Utility"
                ],
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "Session ID",
                        "x-ms-summary": "Session ID",
                        "x-ms-url-encoding": "single"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": [
                                "script"
                            ],
                            "properties": {
                                "script": {
                                    "type": "string",
                                    "description": "JavaScript code to execute in the page context",
                                    "x-ms-summary": "Script"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JavaScript result",
                        "schema": {
                            "$ref": "#/definitions/EvalResponse"
                        }
                    },
                    "404": {
                        "description": "Session not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "SessionStatusResponse": {
            "type": "object",
            "properties": {
                "sessionId": {
                    "type": "string",
                    "description": "Session identifier",
                    "x-ms-summary": "Session ID"
                },
                "url": {
                    "type": "string",
                    "description": "Current page URL",
                    "x-ms-summary": "URL"
                },
                "title": {
                    "type": "string",
                    "description": "Current page title",
                    "x-ms-summary": "Title"
                },
                "createdAt": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Session creation time",
                    "x-ms-summary": "Created At"
                },
                "expiresAt": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Session expiration time",
                    "x-ms-summary": "Expires At"
                },
                "actionCount": {
                    "type": "integer",
                    "description": "Number of actions performed",
                    "x-ms-summary": "Action Count"
                }
            }
        },
...
        "EvalResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string",
                    "description": "Serialized return value of the JavaScript expression",
                    "x-ms-summary": "Result"
                }
            }
        }
    },
...
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "in": "header",
            "name": "X-API-Key"
        }
    },
    "security": [
        {
            "api_key": []
        }
    ],
...
}